Introduce benchmarks for metric overlay
We're kind of getting killed having to continually clone keys and
values in metric maps. This gets especially bad when we start to
integrate additional languages whose memory models require
copying.

This commit adds benchmarks to help pin down the work needed to
repair the situation before it gets out of hand.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>